# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Облік Майна API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/GET/documents/is-realized": {
      "get": {
        "summary": "Check selected documents status",
        "description": "Check selected documents status. If one of them is realized return true",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids of the documents to check",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON that contains boolean value",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "IsPinned",
                  "properties": {
                    "isRealized": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/GET/main-assets-selected": {
      "get": {
        "summary": "Returns list of selected main assets",
        "description": "Returns list of selected main assets. If query array is empty, return empty array",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids of the main assets to fetch.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON list of main assets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/api-get-main-assets-selected.yaml#/MainAsset"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/GET/main-assets": {
      "get": {
        "summary": "Returns list of main assets",
        "description": "Returns list of main assets",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "status or statuses of the main assets to fetch",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON list of main assets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/api-get-main-assets.yaml#/MainAsset"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/GET/main-assets/is-pinned": {
      "get": {
        "summary": "Check selected main assets status",
        "description": "Check selected main assets status. If one of them is pinned return true",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids of the main assets to check",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON that contains boolean value",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "IsPinned",
                  "properties": {
                    "isPinned": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/GET/users": {
      "get": {
        "summary": "Returns list of users",
        "description": "Returns list of all user names",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON list of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/dbModels/user.yaml#/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/PATCH/document/{operation}/realize/{id}": {
      "patch": {
        "summary": "Realize document",
        "description": "Realize created earlier document",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "parameters": [
          {
            "name": "operation",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "string representation of the operation"
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the document to patch"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}